rework the space-time fabric of VisAD displays: use doubles for animation computations, and check for Real.Time or Units convertible with second; use curved texture mapping in Java3D; build partial topologies when missing spatial ruins topology

+++ core/src/visad/trunk/AnimationControl.java
@@ -43,7 +43,7 @@ public interface AnimationControl extends AVControl {
  
   /** set the current step by the value of the RealType
       mapped to Display.Animation */
-  public void setCurrent(float value)
+  public void setCurrent(double value)
          throws VisADException, RemoteException;
  
   /** get the current ordinal step number */
