+++ core/src/visad/trunk/java3d/AnimationControlJ3D.java
@@ -115,7 +115,7 @@ public class AnimationControlJ3D extends AVControlJ3D
     changeControl(true);
   }
  
-  public void setCurrent(float value)
+  public void setCurrent(double value)
          throws VisADException, RemoteException {
     if (animationSet != null) {
       current = animationSet.getIndex(value);
